Addressed all issues
